\name{cdlei-package}
\alias{cdlei-package}
\docType{package}
\title{
Cause-Deleted Life Expectancy Improvement Procedure}

\description{
The concept of cause-deleted life expectancy improvement is statistic designed to quantify the increase inlife expectancy if a certain cause of death is removed.}
\author{
Peter Adamic, Alicja Wolny-Dominiak
Maintainer: <alicja.wolny-dominiak@ue.katowice.pl>
}
\references{
1. Adamic, P. (2015). Life Expectancy Improvement with a Curve Distribution for a cause of death, Australian Journal of Actuarial Practice, 3, 59-70. \cr
2. Adamic, P. (2008). Cause-deleted life expectancy improvement in the presence of
left and right censoring. Belgian Actuarial Bulletin, 8: 17-21. \cr
3. Brown, R.L. (1997). Introduction to the Mathematics of Demography, 3rd ed, Winsted, Connecticut: Actex. \cr
}


